src/ai_networking.py
```python
from src.integrations.linkedin_api import LinkedInAPI
from src.utils import validateNetworkingPreferences

class AINetworking:
    def __init__(self, user_preferences):
        self.user_preferences = user_preferences
        self.linkedin_api = LinkedInAPI()

    def initiateNetworking(self):
        validateNetworkingPreferences(self.user_preferences)
        potential_contacts = self.linkedin_api.fetchLinkedInContacts(self.user_preferences)
        for contact in potential_contacts:
            self.initiateContact(contact)

    def initiateContact(self, contact):
        message = f"Hello {contact['name']}, I came across your profile on LinkedIn and thought we might have some common interests. Would you be open to connecting?"
        self.linkedin_api.sendMessage(contact['id'], message)
        print(f"Contact initiated with {contact['name']}")
```