tests/test_ai_networking.py
```python
import unittest
from src.ai_networking import initiateNetworking, initiateContact
from src.utils import validateNetworkingPreferences

class TestAINetworking(unittest.TestCase):

    def setUp(self):
        self.networking_preferences = {
            'industry': 'Software',
            'location': 'San Francisco',
            'role': 'Engineer'
        }

    def test_validateNetworkingPreferences(self):
        result = validateNetworkingPreferences(self.networking_preferences)
        self.assertTrue(result)

    def test_initiateNetworking(self):
        result = initiateNetworking(self.networking_preferences)
        self.assertEqual(result, 'networkingInitiated')

    def test_initiateContact(self):
        contact = {
            'name': 'John Doe',
            'role': 'Software Engineer',
            'company': 'Tech Corp',
            'location': 'San Francisco'
        }
        result = initiateContact(contact)
        self.assertEqual(result, 'contactInitiated')

if __name__ == '__main__':
    unittest.main()
```